+++ src/UiAdminBase/Contract/Exception/InvalidArgumentThrowableInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception;
+
+/**
+ * Identifies every exception relating to invalid arguments
+ */
+interface InvalidArgumentThrowableInterface extends UiAdminBaseThrowableInterface
+{
+}